requirements.txt
Flask==2.0.1
pytesseract==0.3.10
opencv-python-headless==4.5.5.64
Pillow==9.0.1
deepface==0.0.75
tensorflow==2.11.0
PyMuPDF==1.20.0
pdf2image==1.16.0
gunicorn==20.1.0
numpy==1.22.4